Add Test Case for Flow Error With Imports
Adds a test case for using a custom import in the imports list of a
custom generator. The test itself passes, but running a flow type check
turns up several errors. Also made some tweaks to the test-suite.js file
to allow for running a specific subset of tests from the command line.