Add Support for Flow Utility Types: Option One
This splits MetaType into a union of LiteralType and NonLiteralType (the
names could potentially be improved) that are sentineled by a bool. This
approach has some things I don't like, like the bool sentinel field that
consumers will need to add to all their defined MetaTypes or that
LiteralType has a name field that is just 'literal' (removing it would
be possible, but there's a lot of places that assume MetaType has a name
field).